+++ src/include/patomic/macros/have_twos_compl.h
@@ -1,7 +1,0 @@
-#ifndef PATOMIC_HAVE_TWOS_COMPL
-
-#include <limits.h>
-
-#define PATOMIC_HAVE_TWOS_COMPL (-INT_MAX != INT_MIN)
-
-#endif  /* !PATOMIC_HAVE_TWOS_COMPL */